requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
pymemcache==4.0.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
gunicorn>=20.1,<21.0
